src/test/javascript/spec/app/entities/registration/registration-update.component.spec.ts
/* tslint:disable max-line-length */
import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { Observable, of } from 'rxjs';

import { SbcTestTestModule } from '../../../test.module';
import { RegistrationUpdateComponent } from 'app/entities/registration/registration-update.component';
import { RegistrationService } from 'app/entities/registration/registration.service';
import { Registration } from 'app/shared/model/registration.model';

describe('Component Tests', () => {
    describe('Registration Management Update Component', () => {
        let comp: RegistrationUpdateComponent;
        let fixture: ComponentFixture<RegistrationUpdateComponent>;
        let service: RegistrationService;

        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [SbcTestTestModule],
                declarations: [RegistrationUpdateComponent]
            })
                .overrideTemplate(RegistrationUpdateComponent, '')
                .compileComponents();

            fixture = TestBed.createComponent(RegistrationUpdateComponent);
            comp = fixture.componentInstance;
            service = fixture.debugElement.injector.get(RegistrationService);
        });

        describe('save', () => {
            it(
                'Should call update service on save for existing entity',
                fakeAsync(() => {
                    // GIVEN
                    const entity = new Registration(123);
                    spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
                    comp.registration = entity;
                    // WHEN
                    comp.save();
                    tick(); // simulate async

                    // THEN
                    expect(service.update).toHaveBeenCalledWith(entity);
                    expect(comp.isSaving).toEqual(false);
                })
            );

            it(
                'Should call create service on save for new entity',
                fakeAsync(() => {
                    // GIVEN
                    const entity = new Registration();
                    spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
                    comp.registration = entity;
                    // WHEN
                    comp.save();
                    tick(); // simulate async

                    // THEN
                    expect(service.create).toHaveBeenCalledWith(entity);
                    expect(comp.isSaving).toEqual(false);
                })
            );
        });
    });
});
